removes cameraWebviewController and JsMessageHandler coupling
