--- WkCamera/WkCamera/MetalFilter.swift
@@ -1,32 +1,0 @@
-//
-//  MetalFilter.swift
-//  jscamera
-//
-//  Created by Yves Delacrétaz on 01.07.21.
-//
-
-import Foundation
-import CoreImage
-
-public class MetalFilterSwapBandR: CIFilter {
-
-    private let kernel: CIColorKernel
-
-    var inputImage: CIImage?
-
-    override init() {
-        let url = Bundle.main.url(forResource: "default", withExtension: "metallib")!
-        let data = try! Data(contentsOf: url)
-        kernel = try! CIColorKernel(functionName: "swapRedAndBlueAmount", fromMetalLibraryData: data)
-        super.init()
-    }
-
-    required init?(coder aDecoder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-
-    func outputImage() -> CIImage? {
-        guard let inputImage = inputImage else {return nil}
-        return kernel.apply(extent: inputImage.extent, arguments: [inputImage])
-    }
-}
